+++ pages/SelectorPage/index.js
@@ -1,9 +1,19 @@
 import SportList from "../../components/SportList";
 
-export default function SelectorPage() {
+export default function SelectorPage({
+  sports,
+  selectedSport,
+  onSelectSport,
+  onNewRoll,
+}) {
   return (
     <>
-      <SportList />
+      <SportList
+        sports={sports}
+        selectedSport={selectedSport}
+        onNewRoll={onNewRoll}
+        onSelectSport={onSelectSport}
+      />
     </>
   );
 }
